Move functions and settings endpoints
