+++ src/utils/mime.rs
@@ -1,0 +1,53 @@
+use std::collections::BTreeMap;
+use rocket::http::ContentType;
+
+pub(crate) fn mime_types() -> BTreeMap<String, ContentType> {
+    BTreeMap::from([
+        ("aac".to_string(), ContentType::AAC),
+        ("bin".to_string(), ContentType::Binary),
+        ("bmp".to_string(), ContentType::BMP),
+        ("css".to_string(), ContentType::CSS),
+        ("csv".to_string(), ContentType::CSV),
+        ("flac".to_string(), ContentType::FLAC),
+        ("gif".to_string(), ContentType::GIF),
+        ("gz".to_string(), ContentType::GZIP),
+        ("htm".to_string(), ContentType::HTML),
+        ("html".to_string(), ContentType::HTML),
+        ("ico".to_string(), ContentType::Icon),
+        ("ics".to_string(), ContentType::Calendar),
+        ("jpeg".to_string(), ContentType::JPEG),
+        ("jpg".to_string(), ContentType::JPEG),
+        ("js".to_string(), ContentType::JavaScript),
+        ("md".to_string(), ContentType::Markdown),
+        ("mov".to_string(), ContentType::MOV),
+        ("mp4".to_string(), ContentType::MP4),
+        ("mpeg".to_string(), ContentType::MPEG),
+        ("mpeg4".to_string(), ContentType::MP4),
+        ("mpg".to_string(), ContentType::MPEG),
+        ("ogg".to_string(), ContentType::OGG),
+        ("ogv".to_string(), ContentType::OGG),
+        ("otf".to_string(), ContentType::OTF),
+        ("pdf".to_string(), ContentType::PDF),
+        ("png".to_string(), ContentType::PNG),
+        ("svg".to_string(), ContentType::SVG),
+        ("tar".to_string(), ContentType::TAR),
+        ("tif".to_string(), ContentType::TIFF),
+        ("tiff".to_string(), ContentType::TIFF),
+        ("tsv".to_string(), ContentType::CSV),
+        ("ttf".to_string(), ContentType::TTF),
+        ("txt".to_string(), ContentType::Plain),
+        ("usfm".to_string(), ContentType::Plain),
+        ("usj".to_string(), ContentType::JSON),
+        ("usx".to_string(), ContentType::XML),
+        ("vrs".to_string(), ContentType::Plain),
+        ("wasm".to_string(), ContentType::WASM),
+        ("wav".to_string(), ContentType::WAV),
+        ("weba".to_string(), ContentType::WEBA),
+        ("webm".to_string(), ContentType::WEBM),
+        ("webp".to_string(), ContentType::WEBP),
+        ("woff".to_string(), ContentType::WOFF),
+        ("woff2".to_string(), ContentType::WOFF2),
+        ("xml".to_string(), ContentType::XML),
+        ("zip".to_string(), ContentType::ZIP),
+    ])
+}
